requirements.txt
sqlalchemy
boto3
psycopg2-binary